Rename tasks section folder
