+++ assets/home/home.scss
@@ -5,7 +5,7 @@ $break-medium: 783px; // adminbar goes big
 @import '../shared/styles/variables';
 @import '../shared/styles/wp-admin';
 @import '../css/font-recoleta';
-@import 'tasks/style';
+@import 'tasks-section/style';
 
 @mixin white-box( $r: 2px ) {
 	border-radius: $r;
