+++ assets/home/main.js
@@ -12,7 +12,7 @@ import SenseiProAd from './sections/sensei-pro-ad';
 import Header from './header';
 import { Col, Grid } from './grid';
 import QuickLinks from './sections/quick-links';
-import TasksSection from './tasks';
+import TasksSection from './tasks-section';
 import GetHelp from './sections/get-help';
 import SenseiGuides from './sections/sensei-guides';
 import LatestNews from './sections/latest-news';
