+++ assets/home/tasks-section/progress.scss
